Move relation validations after standardization
