+++ libs/datamodel/core/src/transform/ast_to_dml/db.rs
@@ -100,11 +100,9 @@ impl<'ast> ParserDatabase<'ast> {
 
         // Fourth step: global validations
         attributes::fill_in_default_constraint_names(&mut ctx);
-        attributes::validate_relation_fields(&mut ctx);
 
         // Fifth step: relation inference
         relations::infer_relations(&mut ctx);
-        relations::validate_relations(&mut ctx);
 
         ctx.finish()
     }
